+++ test/metrics/distance/levenshtein.js
@@ -9,6 +9,7 @@ import levenshtein, {limited} from '../../../src/metrics/distance/levenshtein';
 describe('levenshtein', function() {
   const tests = [
     [['b', 'o', 'o', 'k'], ['b', 'a', 'c', 'k'], 2],
+    [['the', 'cat', 'eats', 'mouse'], ['the', 'mouse', 'likes', 'mouse'], 2],
     ['book', 'back', 2],
     ['hello', 'helo', 1],
     ['good sir', 'baal', 8],